Changes on Scorecard design

+++ src/lib/s3-helper.js
@@ -3,10 +3,10 @@ import pnFs from 'pn/fs';
 import path from 'path';
 import uuid from 'node-uuid';
 import AWS from 'aws-sdk';
-import gm from 'gm';
 import util from 'util';
 import svg2png from 'svg2png';
 import Config from '../res/config';
+const exec = require('child_process').exec;
 
 const credentials = new AWS.SharedIniFileCredentials({profile: Config.AWS_PROFILE});
 AWS.config.credentials = credentials;
@@ -21,47 +21,45 @@ export default {
             const pngFilename = `${uuid.v4()}.png`;
             console.log(`filename: ${pngFilename}`);
             const pngFilepath = path.join(__dirname, '..', '..', 'gen', pngFilename);
-            gm(filepath)
-                .density(100, 100)
-                .quality(50)
-                .resize(null, 400)
-                .write(pngFilepath, function (err) {
-                    if (err) {
-                        return console.log(err);
+
+            const cmd = `convert -density 250 ${filepath} ${pngFilepath}`;
+            exec(cmd, (err, stdout, stderr) => {
+                if (err || stderr) {
+                    return console.log(err || stderr);
+                }
+
+                console.log(`image converted, ${pngFilepath}`);
+
+                const body = fs.createReadStream(pngFilepath);
+                const s3obj = new AWS.S3({
+                    params: {
+                        Bucket: Config.CHARTS_BUCKET,
+                        Key: `charts/${pngFilename}`
                     }
+                });
 
-                    console.log(`image converted, ${pngFilepath}`);
+                s3obj.upload({Body: body}).on('httpUploadProgress', evt => {
+                    console.log(`httpUploadProgress: ${evt}`);
+                }).send((err, data) => {
+                    if (err) {
+                        callback(err);
+                    } else {
+                        callback(null, data.Location);
+                    }
+                    console.log(util.inspect(data));
 
-                    const body = fs.createReadStream(pngFilepath);
-                    const s3obj = new AWS.S3({
-                        params: {
-                            Bucket: Config.CHARTS_BUCKET,
-                            Key: `charts/${pngFilename}`
+                    fs.unlink(filepath, err => {
+                        if (!err) {
+                            console.log(`Deleted ${filepath}`);
                         }
                     });
-
-                    s3obj.upload({Body: body}).on('httpUploadProgress', evt => {
-                        console.log(`httpUploadProgress: ${evt}`);
-                    }).send((err, data) => {
-                        if (err) {
-                            callback(err);
-                        } else {
-                            callback(null, data.Location);
+                    fs.unlink(pngFilepath, err => {
+                        if (!err) {
+                            console.log(`Deleted ${pngFilepath}`);
                         }
-                        console.log(util.inspect(data));
-
-                        fs.unlink(filepath, err => {
-                            if (!err) {
-                                console.log(`Deleted ${filepath}`);
-                            }
-                        });
-                        fs.unlink(pngFilepath, err => {
-                            if (!err) {
-                                console.log(`Deleted ${pngFilepath}`);
-                            }
-                        });
                     });
                 });
+            });
         } else {
             const pngFilename = `${uuid.v4()}.png`;
             console.log(`filename: ${pngFilename}`);
